Use unsigned int for variables that use bitwise operations. Fix zstream.cpp where an optional value wasn't being taken by reference (the optional itself was).

+++ include/wx/access.h
@@ -310,7 +310,7 @@ public:
          { return wxAccStatus::NotImplemented; }
 
         // Returns a state constant.
-    virtual wxAccStatus GetState(int WXUNUSED(childId), long* WXUNUSED(state))
+    virtual wxAccStatus GetState(int WXUNUSED(childId), unsigned int* WXUNUSED(state))
          { return wxAccStatus::NotImplemented; }
 
         // Returns a localized string representing the value for the object
